design: systematic UI overhaul — systematic polish pass
- Radius tiers: cards 16px, buttons 12px, inputs 10px, chips 20px
- Shadows: shadow-lg → shadow-sm/md everywhere, remove colored glows
- De-gradient: tracker stats, alerts header, profile avatar all flat now
- Alerts: replaced gradient hero header with clean text header
- Profile: solid indigo avatar, removed decorative blob, tighter type
- Discover card: rounded-2xl, shadow-lg (not xl), tighter buttons
- Login: rounded-2xl card, rounded-lg toggle, smaller hero text
- Splash: smaller logo container, reduced shadow, no bounce animation
- Tracker: all 3 stat cards now white with uniform borders
- Kanban: column rounded-xl, cards rounded-lg
- Base: Button h-12, font-semibold; Input h-12, text-sm

+++ backend/app/api/scholarships.py
@@ -17,7 +17,7 @@ class ScholarshipResponse(BaseModel):
     title: str
     amount: int
     deadline_days: int
-    effort_hours: int
+    effort_hours: float
     win_probability: int
     tags: Optional[str] = None
     description: Optional[str] = None
